requirements.txt
web3>=6.0.0
aiohttp>=3.8.0
numpy>=1.21.0
python-dotenv>=0.19.0
eth-account>=0.5.9